Added bin merging, bin class, bugfix
Added a new bin class used to keep track of bins, their intervals
and marked status. Used it to allow for merging of the adjacent
bins with the same marked status as determined by the chisquare
test.

Fixed a bug which would create an additional bin when an out of
bound index would appear during counting of the support set.

Refactored the code to work with the bin class.